Create ControllerFrame
Controller Frame class -> shows the GUI third tab with a Tkinter Notebook. takes in the data object from the backplane, packs the alarms row, label/table rows and backplane/card data. It also loops through the data object to see if there are more than one backplanes, if so it grabs the second backplane and so on and adds it to a Notebook showing at the bottom of the page with all its corresponding data.